Step 1: Add controls on the map
We add:
* `yaga-zoom-control` to provide zoom buttons
* `yaga-attribution-control` to provide attributions of active layers
* `yaga-scale-control` to provide a scale on the map

We must use an Angular setter to pass the value false (not the string
"false") to the attribution-control directive. An Angular setter is an
HTML attribute that can handle dynamic values from the Angular
environment instead of using static (string) values.

This knowledge we also apply to the `lat`, `lng` and `zoom` attributes
of the map, because it should also use numbers and not their string
representation.