Step 4: Convert simple marker into a location picker
We add:
* a data-model for a location-picker including `lat` and `lng` for the
coordinates and `draggable` for the draggable state of picker.
* the `[draggable]` setter to the marker. We also connect this to our
model (we will use it later interactive).

Changes:
* bind `lat` and `lng` of the marker to the data-model, also. In
addition enhance the Angular setter to a two-way-bound. That means that
changes on the directive will be updated in the data-model and changes
on the data-model will be updated in the directive.
* the popup content will now display the coordinates of the marker via
Angular's interpolation in templates.

Angular provides interpolation in templates (eg.:
`{{ path.within.dataModel }}`). This provides a simple way to output
values from the data-model.

Angular's syntax to provide a binding between the data-model and a
directive in both directions is `[(attributeName)]`. Just to remember,
when you only want to provide values and changes from the data-model
to the directive you use this syntax: `[attributeName]`.