Step 7: Make base-layers configurable over the data-model
We add:
* An array of objects to our data-model describing a base-layer in the
layer-tree.
* `*ngFor` structural-directive to iterate over the baseLayer objects of
the baseLayer array.
* a dynamic tile-layer definition via `*ngFor` and Angular setters

Changes:
* remove static tile-layer definitions

`*ngFor` is a structural directive of Angular. It provides the
possibility to iterate over a list of elements. You can access an
element of the list via a variable similar like you access the
data-model. For example if you have a list of names in a property called
`participants` in your data-model and want to access each name in a
variable called `participant` you can iterate like this:
`<ul><li *ngFor="let participants of data.participants">{{ participant }}</li></ul>`
and get a list of the names of the participants. We do the same with the
layers.