Step 8: Load and display points from a geojson file
We add:
* a marker and iterate over the features of the provided points in a
geojson file.
* using the `getFeatureCollection` method of our app-component to get
the content of a geojson file located in the assets.
* pipe to `async`, because loading content can not be provided
synchronous.
* an OR operator (`||`) to provide an empty array of features as
fallback.

Before we start the tutorial, we already implement the method
`getFeatureCollection`in our app-component. This method provides a way
to serve the content of a geojson file to Angular. The file must be
stored in `src/assets/geojson/` folder. The parameter of the function
must be the filename. For example to get a geojson called
`example.geojson` stored in the geojson folder you have to call
`getFeatureCollection('filename')`. In most cases you will also provide
a pipe (Angular syntax is: ` | `) to `async`. This will wait for the
result of a asynchronous call. Fetching data should not be done in
synchronous way. As long as the data is not available, it will be
represented as value `null`. This is why we provide a fallback value
with the OR-Operator (` || `). This operator will provide the value
given on the left side, excepts it is not available then it will return
the value of the right side. For example if `data.someValue` is not
available you can provide a fallback value with:
`data.someValue || 'no data available''`.